requirements.txt
# API and HTTP
googlemaps==4.10.0
requests==2.31.0
fastapi==0.109.0
uvicorn==0.27.0
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.16.0
snscrape==0.7.0.20240124

# Data Processing
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
spacy==3.7.2
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.36.0
prophet==1.1.5
dask==2024.1.1

# Forecasting and ML
xgboost==2.0.3
lightgbm==4.2.0
optuna==3.5.0
statsmodels==0.14.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
chromadb==0.4.22
neo4j==5.14.1

# Vector Search & LLM
langchain==0.1.0
faiss-cpu==1.7.4
openai==1.12.0
tiktoken==0.5.2
llama-index==0.9.48

# Streaming & Message Queue
kafka-python==2.0.2
confluent-kafka==2.3.0

# Geospatial & Mapping
h3==3.7.6
geopy==2.4.1
geopandas==0.14.1
osmnx==1.9.1
shapely==2.0.2
rtree==1.1.0
networkx==3.2.1
folium==0.15.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
streamlit==1.29.0
plotly==5.18.0

# Communication
twilio==8.12.0
emoji==2.8.0

# ETL & Pipeline
apache-airflow==2.7.3
dask==2024.1.1

# NLP & Text Processing
nltk==3.8.1
gensim==4.3.2
textblob==0.17.1
bertopic==0.15.0
vaderSentiment==3.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.5

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Documentation
sphinx==7.2.6
mkdocs==1.5.3
mkdocs-material==9.5.3